requirements.txt
django==3.1
psycopg2-binary==2.8.5
gunicorn==20.0.4
djangorestframework==3.11.1
pyjwt==1.7.1
